Add static publicPath so additional blocks are found
`--outputDir` puts the build output in some location, but if we want to
split and bundle additional files, webpack doesn't know where to send
the requests.

For example, `import( './edit' ).then( /* … */ )` will produce a file
like `0.js` in the location specified by `outputDir`. However, when the
script is evaluated and the `0.js` chunk is requested, it will only be
found if it happens to live in the directory corresponding to the
current route.

Real case:

```sh
./bin/sdk-cli.js gutenberg client/gutenberg/extensions/presets/jetpack \
--output-dir=/some-wp-site/wp-content/plugins/jetpack/_inc/blocks
```

When we enqueue the script at the correct location, and load
`/wp-admin/post-new.php`, the `0.js` chunk will 404:

`/wp-admin/0.js -> 404`

If we can specify webpack's `output.publicPath` for the location we know
the asset will be served from, the chunk can load correctly.

`/wp-content/plugins/jetpack/_inc/blocks/0.js -> 200`